Break StickyArticleCollection into services
